Fixed projecting of plot extent
